requirements.txt
streamlit==0.89.0
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
simpleai==1.0.0
opencv-python-headless==4.5.3.56
